+++ crates/blockifier/src/transaction/account_transaction.rs
@@ -13,9 +13,8 @@ use crate::execution::contract_class::ContractClass;
 use crate::execution::entry_point::{
     CallEntryPoint, CallType, EntryPointExecutionContext, ExecutionResources,
 };
-use crate::fee::actual_cost::{
-    ActualCost, ActualCostBuilder, FeeCheckReportFields, PostExecutionReport,
-};
+use crate::fee::actual_cost::{ActualCost, ActualCostBuilder};
+use crate::fee::fee_checks::{FeeCheckReportFields, PostExecutionReport};
 use crate::fee::fee_utils::verify_can_pay_max_fee;
 use crate::fee::gas_usage::estimate_minimal_fee;
 use crate::retdata;
